extract constructor machinery into new module
