+++ src/front/wgsl/mod.rs
@@ -4,6 +4,7 @@ Frontend for [WGSL][wgsl] (WebGPU Shading Language).
 [wgsl]: https://gpuweb.github.io/gpuweb/wgsl.html
 */
 
+mod construction;
 mod conv;
 mod lexer;
 mod number_literals;
@@ -496,121 +497,6 @@ impl<'a> Error<'a> {
                 labels: vec![],
                 notes: vec![],
             },
-        }
-    }
-}
-
-/// Represents the type of the constructor
-///
-/// Vectors, Matrices and Arrays can have partial type information
-/// which later gets inferred from the constructor parameters
-enum ConstructorType {
-    Scalar {
-        kind: crate::ScalarKind,
-        width: Bytes,
-    },
-    PartialVector {
-        size: crate::VectorSize,
-    },
-    Vector {
-        size: crate::VectorSize,
-        kind: crate::ScalarKind,
-        width: Bytes,
-    },
-    PartialMatrix {
-        columns: crate::VectorSize,
-        rows: crate::VectorSize,
-    },
-    Matrix {
-        columns: crate::VectorSize,
-        rows: crate::VectorSize,
-        width: Bytes,
-    },
-    PartialArray,
-    Array {
-        base: Handle<crate::Type>,
-        size: crate::ArraySize,
-        stride: u32,
-    },
-    Struct(Handle<crate::Type>),
-}
-
-impl ConstructorType {
-    fn to_type_resolution(&self) -> Option<TypeResolution> {
-        Some(match *self {
-            ConstructorType::Scalar { kind, width } => {
-                TypeResolution::Value(crate::TypeInner::Scalar { kind, width })
-            }
-            ConstructorType::Vector { size, kind, width } => {
-                TypeResolution::Value(crate::TypeInner::Vector { size, kind, width })
-            }
-            ConstructorType::Matrix {
-                columns,
-                rows,
-                width,
-            } => TypeResolution::Value(crate::TypeInner::Matrix {
-                columns,
-                rows,
-                width,
-            }),
-            ConstructorType::Array { base, size, stride } => {
-                TypeResolution::Value(crate::TypeInner::Array { base, size, stride })
-            }
-            ConstructorType::Struct(handle) => TypeResolution::Handle(handle),
-            _ => return None,
-        })
-    }
-}
-
-impl ConstructorType {
-    fn to_error_string(
-        &self,
-        types: &UniqueArena<crate::Type>,
-        constants: &Arena<crate::Constant>,
-    ) -> String {
-        match *self {
-            ConstructorType::Scalar { kind, width } => kind.to_wgsl(width),
-            ConstructorType::PartialVector { size } => {
-                format!("vec{}<?>", size as u32,)
-            }
-            ConstructorType::Vector { size, kind, width } => {
-                format!("vec{}<{}>", size as u32, kind.to_wgsl(width))
-            }
-            ConstructorType::PartialMatrix { columns, rows } => {
-                format!("mat{}x{}<?>", columns as u32, rows as u32,)
-            }
-            ConstructorType::Matrix {
-                columns,
-                rows,
-                width,
-            } => {
-                format!(
-                    "mat{}x{}<{}>",
-                    columns as u32,
-                    rows as u32,
-                    crate::ScalarKind::Float.to_wgsl(width)
-                )
-            }
-            ConstructorType::PartialArray => "array<?, ?>".to_string(),
-            ConstructorType::Array { base, size, .. } => {
-                format!(
-                    "array<{}, {}>",
-                    types[base].name.as_deref().unwrap_or("?"),
-                    match size {
-                        crate::ArraySize::Constant(size) => {
-                            constants[size]
-                                .to_array_length()
-                                .map(|len| len.to_string())
-                                .unwrap_or_else(|| "?".to_string())
-                        }
-                        _ => unreachable!(),
-                    }
-                )
-            }
-            ConstructorType::Struct(handle) => types[handle]
-                .name
-                .clone()
-                .unwrap_or_else(|| "?".to_string()),
         }
     }
 }
@@ -2275,543 +2161,6 @@ impl Parser {
         }))
     }
 
-    fn parse_constructor_type<'a>(
-        &mut self,
-        lexer: &mut Lexer<'a>,
-        word: &'a str,
-        type_arena: &mut UniqueArena<crate::Type>,
-        const_arena: &mut Arena<crate::Constant>,
-    ) -> Result<Option<ConstructorType>, Error<'a>> {
-        if let Some((kind, width)) = conv::get_scalar_type(word) {
-            return Ok(Some(ConstructorType::Scalar { kind, width }));
-        }
-
-        let partial = match word {
-            "vec2" => ConstructorType::PartialVector {
-                size: crate::VectorSize::Bi,
-            },
-            "vec3" => ConstructorType::PartialVector {
-                size: crate::VectorSize::Tri,
-            },
-            "vec4" => ConstructorType::PartialVector {
-                size: crate::VectorSize::Quad,
-            },
-            "mat2x2" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Bi,
-                rows: crate::VectorSize::Bi,
-            },
-            "mat2x3" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Bi,
-                rows: crate::VectorSize::Tri,
-            },
-            "mat2x4" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Bi,
-                rows: crate::VectorSize::Quad,
-            },
-            "mat3x2" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Tri,
-                rows: crate::VectorSize::Bi,
-            },
-            "mat3x3" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Tri,
-                rows: crate::VectorSize::Tri,
-            },
-            "mat3x4" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Tri,
-                rows: crate::VectorSize::Quad,
-            },
-            "mat4x2" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Quad,
-                rows: crate::VectorSize::Bi,
-            },
-            "mat4x3" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Quad,
-                rows: crate::VectorSize::Tri,
-            },
-            "mat4x4" => ConstructorType::PartialMatrix {
-                columns: crate::VectorSize::Quad,
-                rows: crate::VectorSize::Quad,
-            },
-            "array" => ConstructorType::PartialArray,
-            _ => return Ok(None),
-        };
-
-        // parse component type if present
-        match (lexer.peek().0, partial) {
-            (Token::Paren('<'), ConstructorType::PartialVector { size }) => {
-                let (kind, width) = lexer.next_scalar_generic()?;
-                Ok(Some(ConstructorType::Vector { size, kind, width }))
-            }
-            (Token::Paren('<'), ConstructorType::PartialMatrix { columns, rows }) => {
-                let (_, width) = lexer.next_scalar_generic()?;
-                Ok(Some(ConstructorType::Matrix {
-                    columns,
-                    rows,
-                    width,
-                }))
-            }
-            (Token::Paren('<'), ConstructorType::PartialArray) => {
-                lexer.expect_generic_paren('<')?;
-                let base = self
-                    .parse_type_decl(lexer, None, type_arena, const_arena)?
-                    .0;
-                let size = if lexer.skip(Token::Separator(',')) {
-                    let const_handle =
-                        self.parse_const_expression(lexer, type_arena, const_arena)?;
-                    crate::ArraySize::Constant(const_handle)
-                } else {
-                    crate::ArraySize::Dynamic
-                };
-                lexer.expect_generic_paren('>')?;
-
-                let stride = {
-                    self.layouter.update(type_arena, const_arena).unwrap();
-                    self.layouter[base].to_stride()
-                };
-
-                Ok(Some(ConstructorType::Array { base, size, stride }))
-            }
-            (_, partial) => Ok(Some(partial)),
-        }
-    }
-
-    /// Expects [`Scope::PrimaryExpr`] scope on top; if returning Some(_), pops it.
-    fn parse_construction<'a>(
-        &mut self,
-        lexer: &mut Lexer<'a>,
-        type_name: &'a str,
-        type_span: Span,
-        mut ctx: ExpressionContext<'a, '_, '_>,
-    ) -> Result<Option<Handle<crate::Expression>>, Error<'a>> {
-        assert_eq!(
-            self.scopes.last().map(|&(ref scope, _)| scope.clone()),
-            Some(Scope::PrimaryExpr)
-        );
-        let dst_ty = match self.lookup_type.get(type_name) {
-            Some(&handle) => ConstructorType::Struct(handle),
-            None => {
-                match self.parse_constructor_type(lexer, type_name, ctx.types, ctx.constants)? {
-                    Some(inner) => inner,
-                    None => {
-                        match self.parse_type_decl_impl(
-                            lexer,
-                            TypeAttributes::default(),
-                            type_name,
-                            ctx.types,
-                            ctx.constants,
-                        )? {
-                            Some(_) => {
-                                return Err(Error::TypeNotConstructible(type_span));
-                            }
-                            None => return Ok(None),
-                        }
-                    }
-                }
-            }
-        };
-
-        lexer.open_arguments()?;
-
-        let mut components = Vec::new();
-        let mut spans = Vec::new();
-
-        if lexer.peek().0 == Token::Paren(')') {
-            let _ = lexer.next();
-        } else {
-            while components.is_empty() || lexer.next_argument()? {
-                let (component, span) = lexer
-                    .capture_span(|lexer| self.parse_general_expression(lexer, ctx.reborrow()))?;
-                components.push(component);
-                spans.push(span);
-            }
-        }
-
-        enum Components<'a> {
-            None,
-            One {
-                component: Handle<crate::Expression>,
-                span: Span,
-                ty: &'a crate::TypeInner,
-            },
-            Many {
-                components: Vec<Handle<crate::Expression>>,
-                spans: Vec<Span>,
-                first_component_ty: &'a crate::TypeInner,
-            },
-        }
-
-        impl<'a> Components<'a> {
-            fn into_components_vec(self) -> Vec<Handle<crate::Expression>> {
-                match self {
-                    Components::None => vec![],
-                    Components::One { component, .. } => vec![component],
-                    Components::Many { components, .. } => components,
-                }
-            }
-        }
-
-        let components = match *components.as_slice() {
-            [] => Components::None,
-            [component] => {
-                ctx.resolve_type(component)?;
-                Components::One {
-                    component,
-                    span: spans[0].clone(),
-                    ty: ctx.typifier.get(component, ctx.types),
-                }
-            }
-            [component, ..] => {
-                ctx.resolve_type(component)?;
-                Components::Many {
-                    components,
-                    spans,
-                    first_component_ty: ctx.typifier.get(component, ctx.types),
-                }
-            }
-        };
-
-        let expr = match (components, dst_ty) {
-            // Empty constructor
-            (Components::None, dst_ty) => {
-                let ty = match dst_ty.to_type_resolution() {
-                    Some(TypeResolution::Handle(handle)) => handle,
-                    Some(TypeResolution::Value(inner)) => ctx
-                        .types
-                        .insert(crate::Type { name: None, inner }, Default::default()),
-                    None => return Err(Error::TypeNotInferrable(type_span)),
-                };
-
-                return match ctx.create_zero_value_constant(ty) {
-                    Some(constant) => {
-                        let span = self.pop_scope(lexer);
-                        Ok(Some(ctx.interrupt_emitter(
-                            crate::Expression::Constant(constant),
-                            span.into(),
-                        )))
-                    }
-                    None => Err(Error::TypeNotConstructible(type_span)),
-                };
-            }
-
-            // Scalar constructor & conversion (scalar -> scalar)
-            (
-                Components::One {
-                    component,
-                    ty: &crate::TypeInner::Scalar { .. },
-                    ..
-                },
-                ConstructorType::Scalar { kind, width },
-            ) => crate::Expression::As {
-                expr: component,
-                kind,
-                convert: Some(width),
-            },
-
-            // Vector conversion (vector -> vector)
-            (
-                Components::One {
-                    component,
-                    ty: &crate::TypeInner::Vector { size: src_size, .. },
-                    ..
-                },
-                ConstructorType::Vector {
-                    size: dst_size,
-                    kind: dst_kind,
-                    width: dst_width,
-                },
-            ) if dst_size == src_size => crate::Expression::As {
-                expr: component,
-                kind: dst_kind,
-                convert: Some(dst_width),
-            },
-
-            // Matrix conversion (matrix -> matrix)
-            (
-                Components::One {
-                    component,
-                    ty:
-                        &crate::TypeInner::Matrix {
-                            columns: src_columns,
-                            rows: src_rows,
-                            ..
-                        },
-                    ..
-                },
-                ConstructorType::Matrix {
-                    columns: dst_columns,
-                    rows: dst_rows,
-                    width: dst_width,
-                },
-            ) if dst_columns == src_columns && dst_rows == src_rows => crate::Expression::As {
-                expr: component,
-                kind: crate::ScalarKind::Float,
-                convert: Some(dst_width),
-            },
-
-            // Vector constructor (splat) - infer type
-            (
-                Components::One {
-                    component,
-                    ty: &crate::TypeInner::Scalar { .. },
-                    ..
-                },
-                ConstructorType::PartialVector { size },
-            ) => crate::Expression::Splat {
-                size,
-                value: component,
-            },
-
-            // Vector constructor (splat)
-            (
-                Components::One {
-                    component,
-                    ty:
-                        &crate::TypeInner::Scalar {
-                            kind: src_kind,
-                            width: src_width,
-                            ..
-                        },
-                    ..
-                },
-                ConstructorType::Vector {
-                    size,
-                    kind: dst_kind,
-                    width: dst_width,
-                },
-            ) if dst_kind == src_kind || dst_width == src_width => crate::Expression::Splat {
-                size,
-                value: component,
-            },
-
-            // Vector constructor (by elements)
-            (
-                Components::Many {
-                    components,
-                    first_component_ty: &crate::TypeInner::Scalar { kind, width },
-                    ..
-                },
-                ConstructorType::PartialVector { size },
-            )
-            | (
-                Components::Many {
-                    components,
-                    first_component_ty: &crate::TypeInner::Vector { kind, width, .. },
-                    ..
-                },
-                ConstructorType::PartialVector { size },
-            )
-            | (
-                Components::Many {
-                    components,
-                    first_component_ty: &crate::TypeInner::Scalar { .. },
-                    ..
-                },
-                ConstructorType::Vector { size, width, kind },
-            )
-            | (
-                Components::Many {
-                    components,
-                    first_component_ty: &crate::TypeInner::Vector { .. },
-                    ..
-                },
-                ConstructorType::Vector { size, width, kind },
-            ) => {
-                let ty = ctx.types.insert(
-                    crate::Type {
-                        name: None,
-                        inner: crate::TypeInner::Vector { size, kind, width },
-                    },
-                    Default::default(),
-                );
-                crate::Expression::Compose { ty, components }
-            }
-
-            // Matrix constructor (by elements)
-            (
-                Components::Many {
-                    components,
-                    first_component_ty: &crate::TypeInner::Scalar { width, .. },
-                    ..
-                },
-                ConstructorType::PartialMatrix { columns, rows },
-            )
-            | (
-                Components::Many {
-                    components,
-                    first_component_ty: &crate::TypeInner::Scalar { .. },
-                    ..
-                },
-                ConstructorType::Matrix {
-                    columns,
-                    rows,
-                    width,
-                },
-            ) => {
-                let vec_ty = ctx.types.insert(
-                    crate::Type {
-                        name: None,
-                        inner: crate::TypeInner::Vector {
-                            width,
-                            kind: crate::ScalarKind::Float,
-                            size: rows,
-                        },
-                    },
-                    Default::default(),
-                );
-
-                let components = components
-                    .chunks(rows as usize)
-                    .map(|vec_components| {
-                        ctx.expressions.append(
-                            crate::Expression::Compose {
-                                ty: vec_ty,
-                                components: Vec::from(vec_components),
-                            },
-                            crate::Span::default(),
-                        )
-                    })
-                    .collect();
-
-                let ty = ctx.types.insert(
-                    crate::Type {
-                        name: None,
-                        inner: crate::TypeInner::Matrix {
-                            columns,
-                            rows,
-                            width,
-                        },
-                    },
-                    Default::default(),
-                );
-                crate::Expression::Compose { ty, components }
-            }
-
-            // Matrix constructor (by columns)
-            (
-                Components::Many {
-                    components,
-                    first_component_ty: &crate::TypeInner::Vector { width, .. },
-                    ..
-                },
-                ConstructorType::PartialMatrix { columns, rows },
-            )
-            | (
-                Components::Many {
-                    components,
-                    first_component_ty: &crate::TypeInner::Vector { .. },
-                    ..
-                },
-                ConstructorType::Matrix {
-                    columns,
-                    rows,
-                    width,
-                },
-            ) => {
-                let ty = ctx.types.insert(
-                    crate::Type {
-                        name: None,
-                        inner: crate::TypeInner::Matrix {
-                            columns,
-                            rows,
-                            width,
-                        },
-                    },
-                    Default::default(),
-                );
-                crate::Expression::Compose { ty, components }
-            }
-
-            // Array constructor - infer type
-            (components, ConstructorType::PartialArray) => {
-                let components = components.into_components_vec();
-
-                let base = match ctx.typifier[components[0]].clone() {
-                    TypeResolution::Handle(ty) => ty,
-                    TypeResolution::Value(inner) => ctx
-                        .types
-                        .insert(crate::Type { name: None, inner }, Default::default()),
-                };
-
-                let size = crate::Constant {
-                    name: None,
-                    specialization: None,
-                    inner: crate::ConstantInner::Scalar {
-                        width: 4,
-                        value: ScalarValue::Uint(components.len() as u64),
-                    },
-                };
-
-                let inner = crate::TypeInner::Array {
-                    base,
-                    size: crate::ArraySize::Constant(
-                        ctx.constants.append(size, Default::default()),
-                    ),
-                    stride: {
-                        self.layouter.update(ctx.types, ctx.constants).unwrap();
-                        self.layouter[base].to_stride()
-                    },
-                };
-
-                let ty = ctx
-                    .types
-                    .insert(crate::Type { name: None, inner }, Default::default());
-
-                crate::Expression::Compose { ty, components }
-            }
-
-            // Array constructor
-            (components, ConstructorType::Array { base, size, stride }) => {
-                let components = components.into_components_vec();
-                let inner = crate::TypeInner::Array { base, size, stride };
-                let ty = ctx
-                    .types
-                    .insert(crate::Type { name: None, inner }, Default::default());
-                crate::Expression::Compose { ty, components }
-            }
-
-            // Struct constructor
-            (components, ConstructorType::Struct(ty)) => crate::Expression::Compose {
-                ty,
-                components: components.into_components_vec(),
-            },
-
-            // ERRORS
-
-            // Bad conversion (type cast)
-            (
-                Components::One {
-                    span, ty: src_ty, ..
-                },
-                dst_ty,
-            ) => {
-                return Err(Error::BadTypeCast {
-                    span,
-                    from_type: src_ty.to_wgsl(ctx.types, ctx.constants),
-                    to_type: dst_ty.to_error_string(ctx.types, ctx.constants),
-                });
-            }
-
-            // Too many parameters for scalar constructor
-            (Components::Many { spans, .. }, ConstructorType::Scalar { .. }) => {
-                return Err(Error::UnexpectedComponents(Span {
-                    start: spans[1].start,
-                    end: spans.last().unwrap().end,
-                }));
-            }
-
-            // Parameters are of the wrong type for vector or matrix constructor
-            (Components::Many { spans, .. }, ConstructorType::Vector { .. })
-            | (Components::Many { spans, .. }, ConstructorType::Matrix { .. })
-            | (Components::Many { spans, .. }, ConstructorType::PartialVector { .. })
-            | (Components::Many { spans, .. }, ConstructorType::PartialMatrix { .. }) => {
-                return Err(Error::InvalidConstructorComponentType(spans[0].clone(), 0));
-            }
-        };
-
-        let span = NagaSpan::from(self.pop_scope(lexer));
-        Ok(Some(ctx.expressions.append(expr, span)))
-    }
-
     fn parse_const_expression_impl<'a>(
         &mut self,
         first_token_span: TokenSpan<'a>,
@@ -2941,9 +2290,13 @@ impl Parser {
                     TypedExpression::non_reference(expr)
                 } else {
                     let _ = lexer.next();
-                    if let Some(expr) =
-                        self.parse_construction(lexer, word, span.clone(), ctx.reborrow())?
-                    {
+                    if let Some(expr) = construction::parse_construction(
+                        self,
+                        lexer,
+                        word,
+                        span.clone(),
+                        ctx.reborrow(),
+                    )? {
                         TypedExpression::non_reference(expr)
                     } else {
                         return Err(Error::UnknownIdent(span, word));
